pyproject.toml

[tool.poetry]
name = "repo-informalidad"
version = "0.1.0"
description = ""
authors = ["Ollin"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.1"
tqdm = "^4.66.4"
pandas = "^2.2.2"
geopandas = "^1.0.1"
matplotlib = "^3.9.1"
transformers = {extras = ["torch"], version = "^4.43.1"}
datasets = "^2.20.0"
scikit-learn = "^1.5.1"
xgboost = "^2.1.0"
hyperopt = "^0.2.7"
lightgbm = "^4.5.0"
shap = "^0.46.0"
seaborn = "^0.13.2"
altair = "^5.3.0"
imblearn = "^0.0"
scikit-optimize = "^0.10.2"
catboost = "^1.2.5"
numpy = [
    {version = "*", python=">=3.13"},
    {version = "^1.26.0", python=">=3.9,<3.13"},
]
ecomplexity = "^0.5.3"

[tool.poetry.dev-dependencies]
pdoc3 = "^0.10.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"
pandas-stubs = "^2.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# [tool.ruff.lint]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.mypy]
ignore_missing_imports = true
